update examples #12

+++ examples/Ethernet/receive_fastled/receive_fastled.ino
@@ -6,7 +6,7 @@ const IPAddress ip(192, 168, 0, 201);
 uint8_t mac[] = {0x01, 0x23, 0x45, 0x67, 0x89, 0xAB};
 
 ArtnetReceiver artnet;
-uint32_t universe = 1;
+uint8_t universe = 1;  // 0 - 15
 
 // FastLED
 #define NUM_LEDS 1
@@ -21,6 +21,8 @@ void setup() {
 
     Ethernet.begin(mac, ip);
     artnet.begin();
+    // artnet.subscribe_net(0);     // optionally you can change
+    // artnet.subscribe_subnet(0);  // optionally you can change
 
     // if Artnet packet comes to this universe, forward them to fastled directly
     artnet.forward(universe, leds, NUM_LEDS);
